Separate TempUserrepositories into TempUserRepository, TempTokenRepository. Previous code still in TempUserRepositoryX
