+++ backendserver/src/main/java/com/moneytransfer/backendserver/repositories/TempUserRepository.java
@@ -1,98 +1,69 @@
 package com.moneytransfer.backendserver.repositories;
 
-import com.moneytransfer.backendserver.Util;
 import com.moneytransfer.backendserver.objects.User;
 
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Random;
 
+public class TempUserRepository implements UserRepository{
 
-public class TempUserRepository implements UserRepository {
+    private static Map<String, User> userMap = new HashMap<>();
 
-    private static Map<String, User> db = new HashMap<>();
-    private static Map<String, String> tokenMap = new HashMap<>();
-
+    /**
+     * Registers an user to the system
+     *
+     * @param user User object that holds information of the user to be signed up
+     * @return true if sign up was successful, false otherwise
+     */
     @Override
     public boolean save(User user) {
         String userEmail = user.getEmail();
-        if (db.containsKey(userEmail)) {
-            System.out.println("\nUser failed to join moneyTransferApp.");
-            System.out.println("Reason: User Email already in use.");
-            System.out.println("Total number of User: " + db.size());
+        if (userMap.containsKey(userEmail)) {
             return false;
         } else {
-            db.put(userEmail, user);
+            userMap.put(userEmail, user);
         }
-        System.out.println("\nUser successfully joined moneyTransferApp");
-        System.out.println("User Email: " + user.getEmail());
-        System.out.println("User HASH-SALT: " + user.getSalt());
-        System.out.println("User Salted Password: " + user.getSaltedPwHash());
-        System.out.println("Total number of User: " + db.size());
-
         return true;
     }
 
+    /**
+     * remove user with assigned token.
+     *
+     * @param email user email
+     * @return true if successfully saved, else return false.
+     */
     @Override
-    public String login(String email, String password) {
-
-        System.out.println("\nUser Requested Login");
-        System.out.println("User Email: " + email);
-
-        if (db.containsKey(email)) {
-            String temp = password + db.get(email).getSalt();
-            String ret = "0";
-            try {
-                MessageDigest md = MessageDigest.getInstance("SHA-256");
-                md.update(temp.getBytes());
-                byte[] hashedVal = md.digest();
-                ret = Util.byteArrToString(hashedVal);
-            } catch (NoSuchAlgorithmException e) {
-                e.printStackTrace();
-            }
-
-            System.out.println(ret);
-
-            if (!db.get(email).passwordChecker(ret)) {
-                System.out.println("User failed to login: Incorrect password");
-                return null;
-            }
-
-            Random random = new Random();
-            String token = random.ints(48, 123)
-                    .filter(i -> (i <= 57 || i >= 65) && (i <= 90 || i >= 97))
-                    .limit(32)
-                    .collect(StringBuilder::new, StringBuilder::appendCodePoint, StringBuilder::append)
-                    .toString();
-
-            tokenMap.put(token, email);
-            System.out.println("User successfully logged in");
-            return token;
+    public boolean remove(String email) {
+        if (userMap.containsKey(email)) {
+            return userMap.remove(email) != null;
         } else {
-            System.out.println("User failed to login: User not found");
+            return false;
+        }
+    }
+
+    /**
+     * lookup user information from the store using email value,
+     * and return the user object.
+     *
+     * @param email email String.
+     * @return UserObject, if it is not exist then return null.
+     */
+    @Override
+    public User lookupUser(String email) {
+        if (userMap.containsKey(email)) {
+            return userMap.get(email);
+        } else {
             return null;
         }
     }
 
+    /**
+     * return number of user currently have.
+     *
+     * @return size of user stored in store.
+     */
     @Override
-    public boolean logout(String token) {
-        if (tokenMap.containsKey(token)) {
-            tokenMap.remove(token);
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    @Override
-    public User lookupUser(String phoneNumber) {
-        for (User user : db.values()) {
-            if (user.getPhoneNumber().equals(phoneNumber)) {
-                return user;
-            }
-        }
-        return null;
+    public int size() {
+        return userMap.size();
     }
 }
